+++ Makefile
@@ -67,7 +67,9 @@ test: clean lint
 
 .PHONY: license
 license: clean tools
-	$(GO_LICENSER) -d -exclude=protocol/gen-codes -licensor='Apache Software Foundation (ASF)' .
+	$(GO_LICENSER) -d  -exclude=plugins/queue/mmap \
+    -exclude=protocol/gen-codes \
+    -licensor='Apache Software Foundation (ASF)' .
 
 .PHONY: verify
 verify: clean license lint test
