+++ docs/plugins/queue/mmap/README.md
@@ -1,5 +1,5 @@
 # Design
-The mmap-queue is a big, fast and persistent queue based on the memory mapped files. One mmap-queue has a directory to store the whole data. The Queue directory is made up with many segments and 1 meta file. 
+The mmap-queue is a big, fast and persistent queue based on the memory mapped files. One mmap-queue has a directory to store the whole data. The Queue directory is made up with many segments and 1 meta file, which references the design of the [bigqueue](https://github.com/grandecola/bigqueue). 
 
 - Segment: Segment is the real data store center, that provides large-space storage and does not reduce read and write performance as much as possible by using mmap. And we will avoid deleting files by reusing them.
 - Meta: The purpose of meta is to find the data that the consumer needs.
